+++ src/Stack/Package.hs
@@ -1173,7 +1173,9 @@ findCandidate
     -> RIO Ctx (Maybe DotCabalPath)
 findCandidate dirs name = do
     pkg <- asks ctxFile >>= parsePackageNameFromFilePath
-    candidates <- liftIO makeNameCandidates
+    customPreprocessorExts <- view $ configL . to configCustomPreprocessorExts
+    let haskellPreprocessorExts = haskellDefaultPreprocessorExts ++ customPreprocessorExts
+    candidates <- liftIO $ makeNameCandidates haskellPreprocessorExts
     case candidates of
         [candidate] -> return (Just (cons candidate))
         [] -> do
@@ -1193,11 +1195,12 @@ findCandidate dirs name = do
             DotCabalFile{} -> DotCabalFilePath
             DotCabalCFile{} -> DotCabalCFilePath
     paths_pkg pkg = "Paths_" ++ packageNameString pkg
-    makeNameCandidates =
-        liftM (nubOrd . concat) (mapM makeDirCandidates dirs)
-    makeDirCandidates :: Path Abs Dir
+    makeNameCandidates haskellPreprocessorExts =
+        liftM (nubOrd . concat) (mapM (makeDirCandidates haskellPreprocessorExts) dirs)
+    makeDirCandidates :: [Text]
+                      -> Path Abs Dir
                       -> IO [Path Abs File]
-    makeDirCandidates dir =
+    makeDirCandidates haskellPreprocessorExts dir =
         case name of
             DotCabalMain fp -> resolveCandidate dir fp
             DotCabalFile fp -> resolveCandidate dir fp
@@ -1268,8 +1271,8 @@ logPossibilities dirs mn = do
         , flow "but did find:"
         , line <> bulletedList (map pretty possibilities)
         , flow "If you are using a custom preprocessor for this module"
-        , flow "with its own file extension, consider adding the file(s)"
-        , flow "to your .cabal under extra-source-files."
+        , flow "with its own file extension, consider adding the extension"
+        , flow "to the 'custom-preprocessor-extensions' field in stack.yaml."
         ]
   where
     makePossibilities name =